Reorganise option locations in toolbar
